refactor(schemas): migrate Phase 20 endpoint schemas — 8 files, 24 classes (#6042)
- skills_governance.py: migrate GapRequest, ApprovalDecision, GovernanceModeUpdate
  → schemas_agent.py (imports TrustLevel, GovernanceMode from skills.models)
- sandbox.py: migrate SandboxExecuteRequest, SandboxScriptRequest,
  SandboxBatchRequest → schemas_workflows.py
- project.py: migrate PhaseStatusItem, ProjectStatusResponse, ProjectReportResponse
  → schemas_system.py
- phases.py: migrate PhaseEntry, PhasesStatusResponse, ValidationRunResponse
  → schemas_system.py
- redis_service.py: migrate ServiceOperationResponse, ServiceStatusResponse,
  HealthStatusResponse → schemas_system.py
- process_management.py: migrate SpawnRequest, SignalRequest, SpawnResponse
  → schemas_system.py
- playwright.py: migrate TestMessageRequest, FrontendTestRequest,
  SnapshotWithRegionsRequest → schemas_code.py (uses default_factory for ssot_config)
- mcp_registry.py: migrate MCPToolInfo, MCPBridgeInfo, MCPRegistryStats
  → schemas_code.py (dead-code classes — never used in endpoint, preserved
  in schemas for completeness)

Cleanup: removed pre-existing duplicate `from autobot_shared.time_utils import now_utc`
in skills_governance.py; removed dead-code `_ssot_config` and `BaseModel` imports
in playwright.py and other files.
