+++ autobot-backend/api/phases.py
@@ -19,10 +19,14 @@ from typing import List
 
 from autobot_shared.time_utils import utc_timestamp
 from fastapi import APIRouter, HTTPException
-from pydantic import BaseModel
 
 from project_state_manager import get_project_state_manager
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
+from api.schemas_system import (
+    PhaseEntry,
+    PhasesStatusResponse,
+    ValidationRunResponse,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -32,27 +36,6 @@ router = APIRouter()
 # ---------------------------------------------------------------------------
 # Response models
 # ---------------------------------------------------------------------------
-
-
-class PhaseEntry(BaseModel):
-    id: str
-    name: str
-    completion: float
-    is_active: bool
-    is_completed: bool
-
-
-class PhasesStatusResponse(BaseModel):
-    status: str
-    service: str
-    phases: List[PhaseEntry]
-    timestamp: str
-
-
-class ValidationRunResponse(BaseModel):
-    status: str
-    message: str
-    timestamp: str
 
 
 # ---------------------------------------------------------------------------
