+++ autobot-backend/api/sandbox.py
@@ -8,11 +8,15 @@ API endpoints for executing commands in the secure Docker sandbox environment.
 """
 
 import logging
-from typing import Any, Dict, List, Optional
+from typing import Any, Dict, List
 
 from fastapi import APIRouter, Depends, HTTPException
-from pydantic import BaseModel
 
+from api.schemas_workflows import (
+    SandboxBatchRequest,
+    SandboxExecuteRequest,
+    SandboxScriptRequest,
+)
 from auth_middleware import check_admin_permission, get_current_user
 from autobot_shared.error_boundaries import ErrorCategory, with_error_handling
 from constants.network_constants import NetworkConstants
@@ -36,32 +40,6 @@ from api.schemas_code import (
 
 router = APIRouter()
 logger = logging.getLogger(__name__)
-
-
-class SandboxExecuteRequest(BaseModel):
-    command: str
-    security_level: str = "high"  # low, medium, high
-    timeout: int = 300  # 5 minutes default
-    execution_mode: str = "command"  # command, script, batch
-    enable_network: bool = False
-    environment: Optional[dict] = None
-
-
-class SandboxScriptRequest(BaseModel):
-    script_content: str
-    language: str = "bash"  # bash, python, etc.
-    security_level: str = "high"
-    timeout: int = 300
-    enable_network: bool = False
-    environment: Optional[dict] = None
-
-
-class SandboxBatchRequest(BaseModel):
-    commands: List[str]
-    security_level: str = "high"
-    timeout: int = 600  # 10 minutes for batch
-    stop_on_error: bool = True
-    enable_network: bool = False
 
 
 @with_error_handling(
